requirements.txt
python-telegram-bot==20.8
SQLAlchemy>=2.0.25
aiofiles>=23.2.1
python-dotenv>=1.0.1
pytz>=2024.1
nest-asyncio>=1.6.0
alembic>=1.13.1
apscheduler>=3.10.4
loguru>=0.7.2
pymysql>=1.1.0     # MySQL driver (recommended for VPS)